+++ src/api/cleo-api.ts
@@ -5,8 +5,8 @@ const cors = require('cors');
 import {sessionMiddleware} from "./session-config";
 import passport = require("passport");
 require("./passport-config");
-import authRouter from "../user/auth-router";
-import {authGuard} from "../user/auth-guard";
+import authRouter from "../auth/auth-router";
+import {authGuard} from "../auth/auth-guard";
 import journalsRouter from "../journal/journals-router";
 import entriesRouter from "../entry/entries-router";
 import {API_TITLE} from "../environment";
