+++ src/journal/journals-router.ts
@@ -7,7 +7,7 @@ import {
     GetJournalsDTO,
     UpdateJournalDTO
 } from "./journals-dtos";
-import {User} from "../user/user.type";
+import {User} from "../auth/user.type";
 import {sendErrorResponse} from "../utils/send-error-response";
 
 const mapRequest = {
